# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Azure GenAI Backend Lab
  description: Production-minded Azure GenAI backend patterns with Python and FastAPI.
  version: 0.1.0
paths:
...
  /api/v1/chat/stream:
    post:
      tags:
      - streaming
      summary: Stream Chat
      operationId: stream_chat_api_v1_chat_stream_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamingChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StreamingChatRequest:
      properties:
        message:
          type: string
          minLength: 1
          title: Message
        conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Conversation Id
      type: object
      required:
      - message
      title: StreamingChatRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError